feat(be): move roles module into auth/roles lib
