+++ backend/libs/auth/src/index.ts
@@ -1,5 +1,11 @@
 export * from './authenticated-user.class'
+export * from './authenticated-request.interface'
 
 export * from './jwt/jwt-auth.module'
-export * from './jwt/jwt-auth.guard'
 export * from './jwt/jwt.interface'
+export * from './jwt/jwt-auth.guard'
+export * from './jwt/auth-ignore.decorator'
+
+export * from './roles/roles.module'
+export * from './roles/roles.guard'
+export * from './roles/roles.decorator'
