+++ backend/libs/auth/src/roles/roles.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common'
+import { RolesService } from './roles.service'
+
+@Module({
+  providers: [RolesService],
+  exports: [RolesService]
+})
+export class RolesModule {}
